+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Bug fixes:
 * Fix `StringIO#write` to transcode strings with encodings that don't match the `StringIO`'s `external_encoding`. (#2839, @flavorjones)
 * Fix processing of proc rest arguments located at the beginning if there are no actual arguments (#2921, @andrykonchin).
 * Fix `Monitor#exit` to raise `ThreadError` when monitor not owned by the current thread (#2922, @andrykonchin).
+* Fix `MatchData#[]` to support negative `length` argument (#2929, @andrykonchin).
 
 Compatibility: